Changed setup name, small formatting changes
