+++ clientEncryptionToCPKNMigrationSamples/ClientSideKeyVaultKeyToCustomerManagedKey/src/main/java/Migration.java
@@ -37,8 +37,8 @@ public class Migration {
      */
     private static void decryptClientSideKeyVaultKey(String clientSecret, String tenantId, String clientId,
                                                      String storageAccount, String sharedKeyCred, String containerName,
-                                                  String blobName, String blobDecryptName, String keyVaultUrl, String keyname,
-                                                  String keyWrapAlgorithm, String path) {
+                                                     String blobName, String blobDecryptName, String keyVaultUrl, String keyname,
+                                                     String keyWrapAlgorithm, String path) {
         String storageAccountUrl = "https://" + storageAccount + ".blob.core.windows.net";
 
         // Setting encryptedKeyClient with key vault key
