+++ clientEncryptionToCPKNMigrationSamples/ClientSideKeyVaultKeyToCustomerProvidedKey/src/main/java/exampleDataCreator/app.config
