+++ clientEncryptionToCPKNMigrationSamples/ClientSideKeyVaultKeyToCustomerProvidedKey/src/main/java/exampleDataCreator/blobExample.txt
