+++ clientEncryptionToCPKNMigrationSamples/ClientSideKeyVaultKeyToMicrosoftManagedKey/src/main/java/exampleDataCreator/ExampleDataCreator.java
@@ -1,4 +1,4 @@
-package exampleCreation;
+package exampleDataCreator;
 
 import com.azure.core.credential.TokenCredential;
 import com.azure.core.cryptography.AsyncKeyEncryptionKey;
@@ -16,21 +16,20 @@ import com.azure.storage.blob.specialized.cryptography.EncryptedBlobClient;
 import com.azure.storage.blob.specialized.cryptography.EncryptedBlobClientBuilder;
 import com.azure.storage.common.StorageSharedKeyCredential;
 
+import java.io.BufferedReader;
 import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.OutputStream;
+import java.io.InputStreamReader;
 import java.time.OffsetDateTime;
 import java.util.Properties;
-import java.util.Random;
 
 /**
- * Set up by creating client-side encrypted blob in a new container using key vault. Running the ExampleCreation is
- * optional as long as if customer has a client-side encrypted blob ready to be migrated and a local key created for
- * reupload
+ * Set up by creating client-side encrypted blob in a new container using key vault and creates encryption scope for use
+ * in server-side encryption later. Running the ExampleDataCreator is optional as long as if customer has a client-side
+ * encrypted blob ready to be migrated and encryption scope created
  */
-public class ExampleCreation {
+public class ExampleDataCreator {
     /**
      * Creates an Async key for client-side encryption
      */
@@ -43,11 +42,39 @@ public class ExampleCreation {
     }
 
     /**
+     * Creating encryption scope for a Microsoft-managed key
+     */
+    private static void createEncryptionScope(String encryptionScope, String storageAccount, String resourceGroup,
+                                              String subscription) {
+        // Template for command:
+        // az storage account encryption-scope create --name <encryptionScopeName> -s Microsoft.Storage
+        //      --account-name <storageAccountName> -g <resourceGroupName> --subscription <subscriptionName>
+        String command = "az storage account encryption-scope create --name " + encryptionScope + " -s " +
+                "Microsoft.Storage --account-name " + storageAccount + " -g " + resourceGroup + " --subscription "
+                + subscription;
+
+        // Running command line command to create encryption scope for given key
+        Process p = null;
+        try {
+            p = new ProcessBuilder("cmd.exe", "/c", command).start();
+        } catch (IOException ex) {
+            ex.printStackTrace();
+        }
+        // Reading outputs from command line
+        BufferedReader r = new BufferedReader(new InputStreamReader(p.getInputStream()));
+        try {
+            while ((r.readLine()) != null) {}
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    /**
      * Creates example container and blob, then uploads with client-side encryption with key vault
      **/
     public static void setup(String clientSecret, String tenantId, String clientId,
-                             String storageAccount, String sharedKeyCred, String keyVaultUrl, String
-            containerName, String blobName, String keyName, String keyWrapAlgorithm) {
+                             String storageAccount, String sharedKeyCred, String keyVaultUrl,
+                             String containerName, String blobName, String keyName, String keyWrapAlgorithm) {
         String storageAccountUrl = "https://" + storageAccount + ".blob.core.windows.net";
 
         // Creating a BlobServiceClient that allows us to perform container and blob operations, given our storage
@@ -90,7 +117,7 @@ public class ExampleCreation {
 
         // Uploading example blob with client-side encryption
         encryptedBlobClient.uploadFromFile("clientEncryptionToCPKNMigrationSamples\\" +
-                "ClientSideKeyVaultKeyToCustomerProvidedKey\\src\\main\\java\\exampleCreation\\" + blobName, true);
+                "ClientSideKeyVaultKeyToMicrosoftManagedKey\\src\\main\\java\\exampleCreation\\" + blobName, true);
     }
 
     public static void main(String[] args) {
@@ -100,14 +127,16 @@ public class ExampleCreation {
         String storageAccount = null;
         String sharedKeyCred = null;
         String keyVaultUrl = null;
+        String resourceGroup = null;
+        String subscription = null;
         String containerName = null;
         String blobName = null;
         String clientSideEncryptionKeyName = null;
-        String localKeyFileName = null;
+        String encryptionScope = null;
         String keyWrapAlgorithm = null;
 
         String pathToDir = "clientEncryptionToCPKNMigrationSamples\\" +
-                "ClientSideKeyVaultKeyToCustomerProvidedKey\\src\\main\\java\\exampleCreation\\";
+                "ClientSideKeyVaultKeyToMicrosoftManagedKey\\src\\main\\java\\exampleCreation\\";
 
         // Extracting variables from config file
         try (InputStream input = new FileInputStream(pathToDir + "app.config")) {
@@ -122,25 +151,19 @@ public class ExampleCreation {
             containerName = prop.getProperty("containerName");
             blobName = prop.getProperty("blobName");
             clientSideEncryptionKeyName = prop.getProperty("clientSideEncryptionKeyName");
-            localKeyFileName = prop.getProperty("localKeyFileName");
+            encryptionScope = prop.getProperty("encryptionScope");
+            resourceGroup = prop.getProperty("resourceGroup");
+            subscription = prop.getProperty("subscription");
             keyWrapAlgorithm = prop.getProperty("keyWrapAlgorithm");
         } catch (IOException ex) {
             ex.printStackTrace();
         }
 
-        // Creating random local key and storing bytes into local file for later use in decrypting
-        byte[] b = new byte[32];
-        new Random().nextBytes(b);
-        try {
-            OutputStream os = new FileOutputStream(pathToDir + localKeyFileName);
-            os.write(b);
-            os.close();
-        } catch (Exception e) {
-            System.out.println("Exception in saving key: " + e);
-        }
-
+        // Create Microsoft-managed key encryption scope
+        createEncryptionScope(encryptionScope, storageAccount, resourceGroup, subscription);
         // Setup where sample blob is client-side encrypted and uploaded to server
         setup(clientSecret, tenantId, clientId,
-                storageAccount, sharedKeyCred, keyVaultUrl, containerName, blobName, clientSideEncryptionKeyName, keyWrapAlgorithm);
+                storageAccount, sharedKeyCred, keyVaultUrl, containerName, blobName, clientSideEncryptionKeyName,
+                keyWrapAlgorithm);
     }
 }
