+++ clientEncryptionToCPKNMigrationSamples/ClientSideKeyVaultKeyToMicrosoftManagedKey/src/main/java/exampleDataCreator/app.config
