+++ clientEncryptionToCPKNMigrationSamples/ClientSideKeyVaultKeyToMicrosoftManagedKey/src/main/java/exampleDataCreator/blobExample.txt
