+++ clientEncryptionToCPKNMigrationSamples/ClientSideLocalKeyToCustomerManagedKey/src/main/java/Migration.java
@@ -52,8 +52,9 @@ public class Migration {
     /**
      * Downloads client-side encrypted blob, decrypts with a local key, then stores in local file temporarily
      */
-    private static void decryptClientSideLocalKey(String storageAccount, String sharedKeyCred, String containerName, String blobName,
-                                                  String blobDecryptName, AsyncKeyEncryptionKey key, String keyWrapAlgorithm, String path) {
+    private static void decryptClientSideLocalKey(String storageAccount, String sharedKeyCred, String containerName,
+                                                  String blobName, String blobDecryptName, AsyncKeyEncryptionKey key,
+                                                  String keyWrapAlgorithm, String path) {
         String storageAccountUrl = "https://" + storageAccount + ".blob.core.windows.net";
         // Creating encrypted blob client to download blob
         BlobClient blobClient = new BlobClientBuilder()
@@ -138,7 +139,8 @@ public class Migration {
         // Decrypts sample blob then reuploads with server-side encryption using customer-managed keys
         decryptClientSideLocalKey(storageAccount, sharedKeyCred, containerName, blobName, blobNameAfterMigration, key,
                 keyWrapAlgorithm, pathToDir);
-        encryptCustomerManaged(storageAccount, sharedKeyCred, containerName, blobNameAfterMigration, encryptionScope, pathToDir);
+        encryptCustomerManaged(storageAccount, sharedKeyCred, containerName, blobNameAfterMigration, encryptionScope,
+                pathToDir);
         cleanup(blobNameAfterMigration, pathToDir);
     }
 }
