+++ clientEncryptionToCPKNMigrationSamples/ClientSideLocalKeyToCustomerManagedKey/src/main/java/exampleDataCreator/ExampleDataCreator.java
@@ -1,12 +1,13 @@
-package exampleCreation;
+package exampleDataCreator;
 
-import com.azure.core.credential.TokenCredential;
 import com.azure.core.cryptography.AsyncKeyEncryptionKey;
 import com.azure.identity.ClientSecretCredentialBuilder;
 import com.azure.security.keyvault.keys.KeyClient;
 import com.azure.security.keyvault.keys.KeyClientBuilder;
-import com.azure.security.keyvault.keys.cryptography.KeyEncryptionKeyClientBuilder;
+import com.azure.security.keyvault.keys.cryptography.LocalKeyEncryptionKeyClientBuilder;
 import com.azure.security.keyvault.keys.models.CreateRsaKeyOptions;
+import com.azure.security.keyvault.keys.models.JsonWebKey;
+import com.azure.security.keyvault.keys.models.KeyOperation;
 import com.azure.security.keyvault.keys.models.KeyVaultKey;
 import com.azure.storage.blob.BlobClient;
 import com.azure.storage.blob.BlobContainerClient;
@@ -16,42 +17,66 @@ import com.azure.storage.blob.specialized.cryptography.EncryptedBlobClient;
 import com.azure.storage.blob.specialized.cryptography.EncryptedBlobClientBuilder;
 import com.azure.storage.common.StorageSharedKeyCredential;
 
+import javax.crypto.spec.SecretKeySpec;
 import java.io.BufferedReader;
 import java.io.FileInputStream;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.io.OutputStream;
 import java.time.OffsetDateTime;
+import java.util.Arrays;
 import java.util.Properties;
+import java.util.Random;
 
 /**
- * Set up by creating client-side encrypted blob in a new container using key vault and creates encryption scope for use
- * in server-side encryption later. Running the ExampleCreation is optional as long as if customer has a client-side
- * encrypted blob ready to be migrated and encryption scope created
+ * Set up by creating client-side encrypted blob in a new container using an example local key. Saves local key in local
+ * file for decrpytion. Creates key vault key and generates encryption scope for reupload during decryption. Running the
+ * ExampleDataCreator is optional as long as if customer has a client-side encrypted blob ready to be migrated and
+ * encryption scope created
  */
-public class ExampleCreation {
+public class ExampleDataCreator {
     /**
-     * Creates an Async key for client-side encryption
+     * Creates a random, not secure local key to be used in client-side encryption
      */
-    private static AsyncKeyEncryptionKey createAsyncKey(KeyVaultKey key, TokenCredential cred) {
-        AsyncKeyEncryptionKey akek = new KeyEncryptionKeyClientBuilder()
-                .credential(cred)
-                .buildAsyncKeyEncryptionKey(key.getId())
-                .block();
+    private static AsyncKeyEncryptionKey createLocalKey(byte[] byteKey) {
+        JsonWebKey localKey = JsonWebKey.fromAes(new SecretKeySpec(byteKey, "AES"),
+                Arrays.asList(KeyOperation.WRAP_KEY, KeyOperation.UNWRAP_KEY))
+                .setId("my-id");
+        AsyncKeyEncryptionKey akek = new LocalKeyEncryptionKeyClientBuilder()
+                .buildAsyncKeyEncryptionKey(localKey).block();
         return akek;
     }
 
     /**
-     * Creating encryption scope for a Microsoft-managed key
+     * Creating encryption scope that allows access to key from key vault
      */
-    private static void createEncryptionScope(String encryptionScope,
-                                              String storageAccount, String resourceGroup, String subscription) {
+    private static void createKeyVaultEncryptionScope(String clientSecret, String tenantId, String clientId,
+                                                      String keyVaultUrl, String keyName, String encryptionScope,
+                                                      String storageAccount, String resourceGroup, String subscription) {
+        // Creating key client that allows access of key vault
+        KeyClient keyClient = new KeyClientBuilder()
+                .vaultUrl(keyVaultUrl)
+                .credential(new ClientSecretCredentialBuilder()
+                        .clientSecret(clientSecret)
+                        .tenantId(tenantId)
+                        .clientId(clientId)
+                        .build())
+                .buildClient();
+
+        // Create key for use in server-side encryption
+        KeyVaultKey rsaKey = keyClient.createRsaKey(new CreateRsaKeyOptions(keyName)
+                .setExpiresOn(OffsetDateTime.now().plusYears(1))
+                .setKeySize(2048));
+        String keyId = rsaKey.getId();
+
         // Template for command:
-        // az storage account encryption-scope create --name <encryptionScopeName> -s Microsoft.Storage
+        // az storage account encryption-scope create --name <encryptionScopeName> -s Microsoft.KeyVault -u <keyID>
         //      --account-name <storageAccountName> -g <resourceGroupName> --subscription <subscriptionName>
         String command = "az storage account encryption-scope create --name " + encryptionScope + " -s " +
-                "Microsoft.Storage --account-name " + storageAccount + " -g " + resourceGroup + " --subscription "
-                + subscription;
+                "Microsoft.KeyVault -u " + keyId +
+                " --account-name " + storageAccount + " -g " + resourceGroup + " --subscription " + subscription;
 
         // Running command line command to create encryption scope for given key
         Process p = null;
@@ -70,11 +95,10 @@ public class ExampleCreation {
     }
 
     /**
-     * Creates example container and blob, then uploads with client-side encryption with key vault
-     **/
-    public static void setup(String clientSecret, String tenantId, String clientId,
-                             String storageAccount, String sharedKeyCred, String keyVaultUrl, String
-            containerName, String blobName, String keyName, String keyWrapAlgorithm) {
+     * Encrypts sample blob using local key provided and uploads to server
+     */
+    public static void setup(String storageAccount, String sharedKeyCred, String containerName, String blobName,
+                                            AsyncKeyEncryptionKey key, String keyWrapAlgorithm) {
         String storageAccountUrl = "https://" + storageAccount + ".blob.core.windows.net";
 
         // Creating a BlobServiceClient that allows us to perform container and blob operations, given our storage
@@ -90,26 +114,10 @@ public class ExampleCreation {
             blobContainerClient.create();
         }
 
-        // Creating key client that allows access of key vault
-        TokenCredential cred = new ClientSecretCredentialBuilder()
-                .clientSecret(clientSecret)
-                .tenantId(tenantId)
-                .clientId(clientId)
-                .build();
-        KeyClient keyClient = new KeyClientBuilder()
-                .vaultUrl(keyVaultUrl)
-                .credential(cred)
-                .buildClient();
-        // Creating an example RSA key in key vault
-        KeyVaultKey rsaKey = keyClient.createRsaKey(new CreateRsaKeyOptions(keyName)
-                .setExpiresOn(OffsetDateTime.now().plusYears(1))
-                .setKeySize(2048));
-
         // Creating a blob client
         BlobClient blobClient = blobContainerClient.getBlobClient(blobName);
 
-        // Setting encryptedKeyClient with key vault key
-        AsyncKeyEncryptionKey key = createAsyncKey(rsaKey, cred);
+        // Setting encryptedKeyClient
         EncryptedBlobClient encryptedBlobClient = new EncryptedBlobClientBuilder()
                 .key(key, keyWrapAlgorithm)
                 .blobClient(blobClient)
@@ -117,7 +125,7 @@ public class ExampleCreation {
 
         // Uploading example blob with client-side encryption
         encryptedBlobClient.uploadFromFile("clientEncryptionToCPKNMigrationSamples\\" +
-                "ClientSideKeyVaultKeyToMicrosoftManagedKey\\src\\main\\java\\exampleCreation\\" + blobName, true);
+                "ClientSideLocalKeyToCustomerManagedKey\\src\\main\\java\\exampleCreation\\" + blobName, true);
     }
 
     public static void main(String[] args) {
@@ -129,14 +137,15 @@ public class ExampleCreation {
         String keyVaultUrl = null;
         String resourceGroup = null;
         String subscription = null;
+        String clientSideLocalKeyFileName = null;
         String containerName = null;
         String blobName = null;
-        String clientSideEncryptionKeyName = null;
+        String serverSideEncryptionKeyName = null;
         String encryptionScope = null;
         String keyWrapAlgorithm = null;
 
         String pathToDir = "clientEncryptionToCPKNMigrationSamples\\" +
-                "ClientSideKeyVaultKeyToMicrosoftManagedKey\\src\\main\\java\\exampleCreation\\";
+                "ClientSideLocalKeyToCustomerManagedKey\\src\\main\\java\\exampleCreation\\";
 
         // Extracting variables from config file
         try (InputStream input = new FileInputStream(pathToDir + "app.config")) {
@@ -148,21 +157,36 @@ public class ExampleCreation {
             storageAccount = prop.getProperty("storageAccount");
             sharedKeyCred = prop.getProperty("sharedKeyCred");
             keyVaultUrl = prop.getProperty("keyVaultUrl");
-            containerName = prop.getProperty("containerName");
-            blobName = prop.getProperty("blobName");
-            clientSideEncryptionKeyName = prop.getProperty("clientSideEncryptionKeyName");
-            encryptionScope = prop.getProperty("encryptionScope");
             resourceGroup = prop.getProperty("resourceGroup");
             subscription = prop.getProperty("subscription");
+            clientSideLocalKeyFileName = prop.getProperty("clientSideLocalKeyFileName");
+            containerName = prop.getProperty("containerName");
+            blobName = prop.getProperty("blobName");
+            serverSideEncryptionKeyName = prop.getProperty("serverSideEncryptionKeyName");
+            encryptionScope = prop.getProperty("encryptionScope");
             keyWrapAlgorithm = prop.getProperty("keyWrapAlgorithm");
         } catch (IOException ex) {
             ex.printStackTrace();
         }
 
-        // Create Microsoft-managed key encryption scope
-        createEncryptionScope(encryptionScope, storageAccount, resourceGroup, subscription);
+        // Creating random local key and storing bytes into local file for later use in decrypting
+        byte[] b = new byte[32];
+        new Random().nextBytes(b);
+        AsyncKeyEncryptionKey key = createLocalKey(b);
+        try {
+            OutputStream os = new FileOutputStream(pathToDir + clientSideLocalKeyFileName);
+            os.write(b);
+            os.close();
+        } catch (Exception e) {
+            System.out.println("Exception in saving key: " + e);
+        }
+
+        // Create an example encryption scope to access given key that will allow for server-side encryption using
+        // customer-managed keys
+        createKeyVaultEncryptionScope(clientSecret, tenantId, clientId,
+                keyVaultUrl, serverSideEncryptionKeyName, encryptionScope, storageAccount, resourceGroup, subscription);
+
         // Setup where sample blob is client-side encrypted and uploaded to server
-        setup(clientSecret, tenantId, clientId,
-                storageAccount, sharedKeyCred, keyVaultUrl, containerName, blobName, clientSideEncryptionKeyName, keyWrapAlgorithm);
+        setup(storageAccount, sharedKeyCred, containerName, blobName, key, keyWrapAlgorithm);
     }
 }
