+++ clientEncryptionToCPKNMigrationSamples/ClientSideLocalKeyToCustomerManagedKey/src/main/java/exampleDataCreator/app.config
