+++ clientEncryptionToCPKNMigrationSamples/ClientSideLocalKeyToCustomerManagedKey/src/main/java/exampleDataCreator/blobExample.txt
