+++ clientEncryptionToCPKNMigrationSamples/ClientSideLocalKeyToCustomerProvidedKey/src/main/java/Migration.java
@@ -53,8 +53,9 @@ public class Migration {
     /**
      * Downloads client-side encrypted blob, decrypts with local key, then stores in local file temporarily
      */
-    private static void decryptClientSideLocalKey(String storageAccount, String sharedKeyCred, String containerName, String blobName,
-                                                  String blobDecryptName, AsyncKeyEncryptionKey key, String keyWrapAlgorithm, String path) {
+    private static void decryptClientSideLocalKey(String storageAccount, String sharedKeyCred, String containerName,
+                                                  String blobName, String blobDecryptName, AsyncKeyEncryptionKey key,
+                                                  String keyWrapAlgorithm, String path) {
         String storageAccountUrl = "https://" + storageAccount + ".blob.core.windows.net";
 
         // Creating encrypted blob client to download blob
@@ -140,7 +141,8 @@ public class Migration {
         // Decrypts sample blob then reuploads with server-side encryption using customer-provided keys
         decryptClientSideLocalKey(storageAccount, sharedKeyCred, containerName, blobName, blobNameAfterMigration, key,
                 keyWrapAlgorithm, pathToDir);
-        encryptCustomerProvided(storageAccount, sharedKeyCred, containerName, blobNameAfterMigration, new CustomerProvidedKey(serverSideKey), pathToDir);
+        encryptCustomerProvided(storageAccount, sharedKeyCred, containerName, blobNameAfterMigration,
+                new CustomerProvidedKey(serverSideKey), pathToDir);
         cleanup(blobNameAfterMigration, pathToDir);
     }
 }
