+++ clientEncryptionToCPKNMigrationSamples/ClientSideLocalKeyToCustomerProvidedKey/src/main/java/README.md
@@ -1,6 +1,6 @@
 ## Additional Setup for Client-Side Encryption with Local Key to Server-Side Encryption with Customer-Provided Keys
 This sample will show how client-side encryption works with local keys and upload an example blob (blobExample.txt found
-in exampleCreation folder) into a newly generated container in the provided storage account linked to an Azure subscription. Then,
+in exampleDataCreator folder) into a newly generated container in the provided storage account linked to an Azure subscription. Then,
 the uploaded blob will be downloaded, decrypted, then reuploaded into the same container with server-side encryption
 using customer-provided keys. Both encryptions will use the same, insecure local key. 
 
@@ -11,11 +11,11 @@ and Maven. Must have an [Azure subscription](https://azure.microsoft.com/en-us/f
 [create a storage account](https://docs.microsoft.com/en-us/azure/storage/common/storage-account-create?tabs=azure-portal).
 
 #### Code Sample Specific Setup
-Requires modification of app.config file in exampleCreation folder. Must add values to variables listed below:
+Requires modification of app.config file in exampleDataCreator folder. Must add values to variables listed below:
  * *sharedKeyCred* (for storage account)
  * *storageAccount*
  
-Requires the following variables in same app.config file, should not be changed if using given exampleCreation.
+Requires the following variables in same app.config file, should not be changed if using given exampleDataCreator.
   * *clientSideLocalKeyFileName* (stores byte array that represents local key)
   * *serverSideLocalKeyFileName* (stores byte array that represents local key)
   * *containerName*
@@ -24,6 +24,6 @@ Requires the following variables in same app.config file, should not be changed 
   
 #### How to use files
 First, ensure to follow setup as described above and that the app.config file is filled out completely. Variables that
-have already been filled out are left as is if ExampleCreation.java will be run. Running the ExampleCreation is optional as 
+have already been filled out are left as is if ExampleDataCreator.java will be run. Running the ExampleDataCreator is optional as 
 long as if customer has a client-side encrypted blob ready to be migrated and a local key prepared for reupload. After
- ExampleCreation is run, run Migration to perform migration to server-side encryption. 
+ ExampleDataCreator is run, run Migration to perform migration to server-side encryption.
