+++ clientEncryptionToCPKNMigrationSamples/ClientSideLocalKeyToCustomerProvidedKey/src/main/java/exampleDataCreator/app.config
