+++ clientEncryptionToCPKNMigrationSamples/ClientSideLocalKeyToCustomerProvidedKey/src/main/java/exampleDataCreator/blobExample.txt
