+++ clientEncryptionToCPKNMigrationSamples/ClientSideLocalKeyToMicrosoftManagedKey/src/main/java/exampleDataCreator/ExampleDataCreator.java
@@ -1,4 +1,4 @@
-package exampleCreation;
+package exampleDataCreator;
 
 import com.azure.core.cryptography.AsyncKeyEncryptionKey;
 import com.azure.security.keyvault.keys.cryptography.LocalKeyEncryptionKeyClientBuilder;
@@ -27,10 +27,10 @@ import java.util.Random;
 /**
  * Set up by creating client-side encrypted blob in a new container using an example local key and creates encryption
  * scope for use in server-side encryption later. Saves local key in local file for decryption. Running the
- * ExampleCreation is optional as long as if customer has a client-side encrypted blob ready to be migrated and
+ * ExampleDataCreator is optional as long as if customer has a client-side encrypted blob ready to be migrated and
  * encryption scope created
  */
-public class ExampleCreation {
+public class ExampleDataCreator {
     /**
      * Creates a random, not secure local key to be used in client-side encryption
      */
