+++ clientEncryptionToCPKNMigrationSamples/ClientSideLocalKeyToMicrosoftManagedKey/src/main/java/exampleDataCreator/app.config
