+++ clientEncryptionToCPKNMigrationSamples/ClientSideLocalKeyToMicrosoftManagedKey/src/main/java/exampleDataCreator/blobExample.txt
